finished poses CRUD, authorisation, db integration

+++ routes/web.php
@@ -1,14 +1,23 @@
 <?php
 
-$router->get('poses', 'PoseController@index');
-$router->get('poses/{id}', 'PoseController@show');
-$router->delete('poses/{id}', 'PoseController@destroy');
-$router->post('poses', 'PoseController@post');
-$router->put('poses/{id}', 'PoseController@update');
+//Authorised
+$router->group(['middleware' => 'auth'], function ($router) {
+    //Poses
+    $router->get('poses', 'PoseController@index');//admin
+    $router->get('poses/pending', 'PoseController@getPendingPoses');//admin
+    $router->delete('poses/{id}', 'PoseController@destroy');//admin
+    $router->post('poses', 'PoseController@post');//user (for confirmation)
+    $router->put('poses/{id}', 'PoseController@update');//admin
+});
+
+//Poses
+$router->get('poses/{id}', 'PoseController@show');//all
+
+//Sequence
+$router->post('sequence/generate', 'SequenceController@generate');//all
+
+//Authentication
+$router->post('login', 'AuthenticationController@login');//all
+$router->post('register', 'AuthenticationController@register');//all
 
 
-/**
- * TODO
- * CRUD for positions
- * Endpoint to fetch positions by difficulty
- */
